Implement JWT-based authentication and authorization with user signup, login, and token refresh functionality.

+++ airBnbApp/src/main/java/com/portfolio/projects/airBnbApp/service/impl/HotelServiceImpl.java
@@ -6,7 +6,9 @@ import com.portfolio.projects.airBnbApp.dto.HotelInfoRequestDto;
 import com.portfolio.projects.airBnbApp.dto.RoomDto;
 import com.portfolio.projects.airBnbApp.entity.Hotel;
 import com.portfolio.projects.airBnbApp.entity.Room;
+import com.portfolio.projects.airBnbApp.entity.User;
 import com.portfolio.projects.airBnbApp.exception.ResourceNotFoundException;
+import com.portfolio.projects.airBnbApp.exception.UnAuthorisedException;
 import com.portfolio.projects.airBnbApp.repository.HotelRepository;
 import com.portfolio.projects.airBnbApp.repository.RoomRepository;
 import com.portfolio.projects.airBnbApp.service.HotelService;
@@ -14,6 +16,7 @@ import com.portfolio.projects.airBnbApp.service.InventoryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,78 +27,103 @@ import java.util.stream.Collectors;
 @Service
 @Slf4j
 @RequiredArgsConstructor
-public class HotelServiceImpl implements HotelService {
-    private final RoomRepository roomRepository;
+public class HotelServiceImpl implements HotelService{
 
     private final HotelRepository hotelRepository;
     private final ModelMapper modelMapper;
     private final InventoryService inventoryService;
-    private final RoomServiceImpl roomService;
+    private final RoomRepository roomRepository;
 
     @Override
     public HotelDto createNewHotel(HotelDto hotelDto) {
-        log.info("Creating new hotel with name: {}", hotelDto.getName());
+        log.info("Creating a new hotel with name: {}", hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto, Hotel.class);
-        hotel.setActive(false); // New hotels are inactive by default
+        hotel.setActive(false);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel = hotelRepository.save(hotel);
-        log.info("Hotel created with ID: {}", hotel.getId());
+        log.info("Created a new hotel with ID: {}", hotelDto.getId());
         return modelMapper.map(hotel, HotelDto.class);
     }
 
     @Override
     public HotelDto getHotelById(Long id) {
-        log.info("Fetching hotel with ID: {}", id);
-        Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + id));
-        log.info("Hotel found: {}", hotel.getName());
+        log.info("Getting the hotel with ID: {}", id);
+        Hotel hotel = hotelRepository
+                .findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+
         return modelMapper.map(hotel, HotelDto.class);
     }
 
     @Override
     public HotelDto updateHotelById(Long id, HotelDto hotelDto) {
         log.info("Updating the hotel with ID: {}", id);
-        Hotel hotel = hotelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + id));
+        Hotel hotel = hotelRepository
+                .findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
-        log.info("Hotel updated with ID: {}", hotel.getId());
         return modelMapper.map(hotel, HotelDto.class);
     }
 
     @Override
     @Transactional
     public void deleteHotelById(Long id) {
-        Hotel hotel = hotelRepository.
-                findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + id));
-        log.info("Deleting the hotel with ID: {}", id);
+        Hotel hotel = hotelRepository
+                .findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
 
 
-        for (Room room: hotel.getRooms()){
+        for(Room room: hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
         }
         hotelRepository.deleteById(id);
-        log.info("Hotel deleted with ID: {}", id);
     }
 
     @Override
     @Transactional
     public void activateHotel(Long hotelId) {
         log.info("Activating the hotel with ID: {}", hotelId);
-        Hotel hotel = hotelRepository.
-                findById(hotelId)
-                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + hotelId));
+        Hotel hotel = hotelRepository
+                .findById(hotelId)
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
+
         hotel.setActive(true);
-        hotelRepository.save(hotel);
-        log.info("Hotel activated with ID: {}", hotel.getId());
 
-        //assuming only do it once
-        for (Room room: hotel.getRooms()){
+        // assuming only do it once
+        for(Room room: hotel.getRooms()) {
             inventoryService.initializeRoomForAYear(room);
-
         }
     }
 
+    //    public method
     @Override
     public HotelInfoDto getHotelInfoById(Long hotelId) {
         Hotel hotel = hotelRepository
@@ -109,4 +137,6 @@ public class HotelServiceImpl implements HotelService {
 
         return new HotelInfoDto(modelMapper.map(hotel, HotelDto.class), rooms);
     }
+
+
 }
