+++ airBnbApp/src/main/java/com/portfolio/projects/airBnbApp/service/impl/RoomServiceImpl.java
@@ -3,13 +3,17 @@ package com.portfolio.projects.airBnbApp.service.impl;
 import com.portfolio.projects.airBnbApp.dto.RoomDto;
 import com.portfolio.projects.airBnbApp.entity.Hotel;
 import com.portfolio.projects.airBnbApp.entity.Room;
+import com.portfolio.projects.airBnbApp.entity.User;
 import com.portfolio.projects.airBnbApp.exception.ResourceNotFoundException;
+import com.portfolio.projects.airBnbApp.exception.UnAuthorisedException;
 import com.portfolio.projects.airBnbApp.repository.HotelRepository;
 import com.portfolio.projects.airBnbApp.repository.RoomRepository;
+import com.portfolio.projects.airBnbApp.service.InventoryService;
 import com.portfolio.projects.airBnbApp.service.RoomService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,65 +23,77 @@ import java.util.stream.Collectors;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class RoomServiceImpl implements RoomService {
+public class RoomServiceImpl implements RoomService{
 
     private final RoomRepository roomRepository;
     private final HotelRepository hotelRepository;
+    private final InventoryService inventoryService;
     private final ModelMapper modelMapper;
-    private final InventoryServiceImpl inventoryService;
+
     @Override
     public RoomDto createNewRoom(Long hotelId, RoomDto roomDto) {
         log.info("Creating a new room in hotel with ID: {}", hotelId);
+        Hotel hotel = hotelRepository
+                .findById(hotelId)
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
 
-        // Verify that the hotel exists
-        Hotel hotel = hotelRepository.
-                findById(hotelId)
-                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + hotelId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
+
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
-        log.info("Room created with ID: {}", room.getId());
-        if(hotel.getActive()){
+
+        if (hotel.getActive()) {
             inventoryService.initializeRoomForAYear(room);
-            log.info("Initialized inventory for room ID: {} for one year", room.getId());
         }
+
         return modelMapper.map(room, RoomDto.class);
     }
 
     @Override
     public List<RoomDto> getAllRoomsInHotel(Long hotelId) {
-        log.info("Fetching all rooms in hotel with ID: {}", hotelId);
-        Hotel hotel = hotelRepository.
-                findById(hotelId)
-                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: " + hotelId));
-        log.info("Found {} rooms in hotel with ID: {}", hotel.getRooms().size(), hotelId);
+        log.info("Getting all rooms in hotel with ID: {}", hotelId);
+        Hotel hotel = hotelRepository
+                .findById(hotelId)
+                .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with ID: "+hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
+
         return hotel.getRooms()
                 .stream()
-                .map(room -> modelMapper.map(room, RoomDto.class))
+                .map((element) -> modelMapper.map(element, RoomDto.class))
                 .collect(Collectors.toList());
     }
 
     @Override
     public RoomDto getRoomById(Long roomId) {
-        log.info("Fetching room with ID: {}", roomId);
-        Room room = roomRepository.
-                findById(roomId)
-                .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: " + roomId));
-        log.info("Room found with ID: {}", roomId);
+        log.info("Getting the room with ID: {}", roomId);
+        Room room = roomRepository
+                .findById(roomId)
+                .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: "+roomId));
         return modelMapper.map(room, RoomDto.class);
     }
 
     @Transactional
     @Override
     public void deleteRoomById(Long roomId) {
-        log.info("Deleting room with ID: {}", roomId);
+        log.info("Deleting the room with ID: {}", roomId);
+        Room room = roomRepository
+                .findById(roomId)
+                .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: "+roomId));
 
-        Room room = roomRepository.
-                findById(roomId)
-                .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: " + roomId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())) {
+            throw new UnAuthorisedException("This user does not own this room with id: "+roomId);
+        }
 
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
-        log.info("Room deleted with ID: {}", roomId);
     }
 }
